KAFKA-18707: Move KRaftMetadataCachePublisher to metadata module
